Delete old rspec files to make way for downgraded version.
In order to have sensible RSpec tests that are understandable by the
majority of the group, we decided to downgrade to RSpec 2.14.1. This
caused pretty severe problems when we didn't delete the files

`uninitialized constant ActiveSupport::AutoLoad (NameError)`

so we decided that deleting and readding the files was the best way to
go.